+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Partially revert length changes to `SpectrumDataset` and `AnnotatedSpectrumDataset`. We removed `__len__` from both due to problems with PyTorch Lightning compatibility.
-- Simplify dataset code by removing redudancy with `lance.pytorch.LanceDatset`.
+- Simplify dataset code by removing redundancy with `lance.pytorch.LanceDatset`.
+- Improved warning message for skipped spectra.
 
 ## [v0.4.3]