+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.4.4]
+
+### Changed
+- Partially revert length changes to `SpectrumDataset` and `AnnotatedSpectrumDataset`. We removed `__len__` from both due to problems with PyTorch Lightning compatibility.
+- Simplify dataset code by removing redundancy with `lance.pytorch.LanceDatset`.
+- Improved warning message for skipped spectra.
+
 ## [v0.4.3]
 
 ### Changed